sdrad_get_stack offset bug fix

+++ src/sdrad_api.h
@@ -185,7 +185,7 @@ int32_t sdrad_call(uint64_t udi, int32_t (*funcptr)(),  void *argument,
  * @param udi 
  * @return stack address
  */
-long sdrad_get_stack_offset(uint64_t udi); 
+intptr_t sdrad_get_stack_offset(uint64_t udi); 
 
 /**
  * @brief RUST sandbox! can set the stack-offset 
@@ -193,5 +193,5 @@ long sdrad_get_stack_offset(uint64_t udi);
  * @param udi 
  * @return stack address
  */
-long sdrad_set_stack_offset(uint64_t udi, long rsp); 
+int64_t sdrad_set_stack_offset(uint64_t udi, uintptr_t rsp);
